+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Restored 2x2 bread and paper recipes
 * Added missing terracotta firing recipe #82
 * Added mixin blacklisting
+* Added Bumblezone compatibility for sugar block when touching water
 
 ## 2.3.2
 * Fixes #68, #72, #61, #59